requirements.txt
cassandra-driver==3.25.0
click==8.1.3
colorama==0.4.4
geomet==0.2.1.post1
packaging==21.3
ply==3.11
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQt5Designer==5.14.1
sip==6.6.1
six==1.16.0
toml==0.10.2
